Fix reboot handling for powershell4.
Was only including the reboot handler for immediate reboots but it is also
needed for delayed reboots.

Added the /norestart and /quiet flags back. Both are needed to reboot
'gracefully' in a way visible to chef.

Added an extra return code which seems to be thrown when a reboot is required
(3010).

In windows_reboot added ruby_block to raise an exception if an immediate reboot
is required. It does nothing otherwise.

Unless no_reboot is set, the reboot will still happen at the end of the chef
run with no exception.